2016 day09: read the repetition count as a Long
`times` feeds a Long multiplication in both parts, so Day09 called
`.toLong()` on it at each of the two use sites. BaseEntity parses Long
directly now, so the entity declares the type it is actually used as.

+++ advent-2016/src/main/kotlin/aoc/year2016/Day09.kt
@@ -61,7 +61,7 @@ class Day09 : AoCFileInput<List<Decompress>, Long>() {
 
                 val dataStartPos = decompress.range.last + 1
                 val dataEndPos = dataStartPos + decompress.num
-                setRepetitionCounts(multiArray, dataStartPos, dataEndPos, decompress.times.toLong())
+                setRepetitionCounts(multiArray, dataStartPos, dataEndPos, decompress.times)
             }
         }
 
@@ -107,7 +107,7 @@ class Day09 : AoCFileInput<List<Decompress>, Long>() {
 
             val dataStartPos = decompress.range.last + 1
             val dataEndPos = dataStartPos + decompress.num
-            multiplyRepetitionCounts(multiArray, dataStartPos, dataEndPos, decompress.times.toLong())
+            multiplyRepetitionCounts(multiArray, dataStartPos, dataEndPos, decompress.times)
         }
 
         return multiArray.sum()
