+++ advent-2016/src/main/kotlin/aoc/year2016/entity/Decompress.kt
@@ -8,5 +8,6 @@ import aoc.ksp.MatchPart
 data class Decompress(
     @FromMatch(MatchPart.RANGE) val range: IntRange,
     val num: Int,
-    val times: Int,
+    // Long because part two multiplies the repetition counts together, which overflows an Int
+    val times: Long,
 )
